refactor: reuse a simple function to do the root key check
