+++ apps/api/src/routes/legacy_apis_createApi.ts
@@ -1,8 +1,9 @@
-import { db, keyService } from "@/pkg/global";
+import { db } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 
-import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
+import { rootKeyAuth } from "@/pkg/auth/root_key";
+import { openApiErrorResponses } from "@/pkg/errors";
 import { KeyAuth, schema } from "@unkey/db";
 import { newId } from "@unkey/id";
 
@@ -52,26 +53,12 @@ export type LegacyApisCreateApiResponse = z.infer<
 
 export const registerLegacyApisCreateApi = (app: App) =>
   app.openapi(route, async (c) => {
-    const authorization = c.req.header("authorization")?.replace("Bearer ", "");
-    if (!authorization) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "key required" });
-    }
-    const rootKey = await keyService.verifyKey(c, { key: authorization });
-    if (rootKey.error) {
-      throw new UnkeyApiError({ code: "INTERNAL_SERVER_ERROR", message: rootKey.error.message });
-    }
-    if (!rootKey.value.valid) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "the root key is not valid" });
-    }
-    if (!rootKey.value.isRootKey) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "root key required" });
-    }
-
+    const auth = await rootKeyAuth(c);
     const { name } = c.req.valid("json");
 
     const keyAuth: KeyAuth = {
       id: newId("keyAuth"),
-      workspaceId: rootKey.value.authorizedWorkspaceId,
+      workspaceId: auth.authorizedWorkspaceId,
       createdAt: new Date(),
       deletedAt: null,
     };
@@ -81,8 +68,8 @@ export const registerLegacyApisCreateApi = (app: App) =>
      * Set up an api for production
      */
     const apiId = newId("api");
-    const authorizedWorkspaceId = rootKey.value.authorizedWorkspaceId;
-    const rootKeyId = rootKey.value.key.id;
+    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
+    const rootKeyId = auth.key.id;
     await db.transaction(async (tx) => {
       await tx.insert(schema.apis).values({
         id: apiId,
