+++ apps/api/src/routes/legacy_keys_deleteKey.ts
@@ -1,8 +1,9 @@
-import { cache, db, keyService } from "@/pkg/global";
+import { cache, db } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 import { schema } from "@unkey/db";
 
+import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
 import { newId } from "@unkey/id";
 import { eq } from "drizzle-orm";
@@ -39,20 +40,7 @@ export type LegacyKeysDeleteKeyResponse = z.infer<
 
 export const registerLegacyKeysDelete = (app: App) =>
   app.openapi(route, async (c) => {
-    const authorization = c.req.header("authorization")?.replace("Bearer ", "");
-    if (!authorization) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "key required" });
-    }
-    const rootKey = await keyService.verifyKey(c, { key: authorization });
-    if (rootKey.error) {
-      throw new UnkeyApiError({ code: "INTERNAL_SERVER_ERROR", message: rootKey.error.message });
-    }
-    if (!rootKey.value.valid) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "the root key is not valid" });
-    }
-    if (!rootKey.value.isRootKey) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "root key required" });
-    }
+    const auth = await rootKeyAuth(c);
 
     const { keyId } = c.req.param();
 
@@ -78,11 +66,11 @@ export const registerLegacyKeysDelete = (app: App) =>
       };
     });
 
-    if (!data || data.key.workspaceId !== rootKey.value.authorizedWorkspaceId) {
+    if (!data || data.key.workspaceId !== auth.authorizedWorkspaceId) {
       throw new UnkeyApiError({ code: "NOT_FOUND", message: `key ${keyId} not found` });
     }
 
-    const authorizedWorkspaceId = rootKey.value.authorizedWorkspaceId;
+    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
 
     await db.transaction(async (tx) => {
       await tx
@@ -96,7 +84,7 @@ export const registerLegacyKeysDelete = (app: App) =>
         time: new Date(),
         workspaceId: authorizedWorkspaceId,
         actorType: "key",
-        actorId: rootKey.value.key!.id,
+        actorId: auth.key!.id,
         event: "key.delete",
         description: `Key ${data.key.id} deleted`,
         apiId: data.api.id,
