+++ apps/api/src/routes/legacy_keys_updateKey.ts
@@ -1,7 +1,8 @@
-import { db, keyService, usageLimiter } from "@/pkg/global";
+import { db, usageLimiter } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 
+import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
 import { schema } from "@unkey/db";
 import { newId } from "@unkey/id";
@@ -119,34 +120,8 @@ export type LegacyKeysUpdateKeyResponse = z.infer<
 
 export const registerLegacyKeysUpdate = (app: App) =>
   app.openapi(route, async (c) => {
-    const authorization = c.req.header("authorization")?.replace("Bearer ", "");
-    if (!authorization) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "key required",
-      });
-    }
+    const auth = await rootKeyAuth(c);
 
-    // Get root key and check for API errors
-    const rootKey = await keyService.verifyKey(c, { key: authorization });
-    if (rootKey.error) {
-      throw new UnkeyApiError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: rootKey.error.message,
-      });
-    }
-    if (!rootKey.value.valid) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "the root key is not valid",
-      });
-    }
-    if (!rootKey.value.isRootKey) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "root key required",
-      });
-    }
     const keyId = c.req.param("keyId");
     const req = c.req.valid("json");
 
@@ -154,15 +129,15 @@ export const registerLegacyKeysUpdate = (app: App) =>
       where: (table, { eq }) => eq(table.id, keyId),
     });
 
-    if (!key || key.workspaceId !== rootKey.value.authorizedWorkspaceId) {
+    if (!key || key.workspaceId !== auth.authorizedWorkspaceId) {
       throw new UnkeyApiError({
         code: "NOT_FOUND",
         message: `key ${keyId} not found`,
       });
     }
 
-    const authorizedWorkspaceId = rootKey.value.authorizedWorkspaceId;
-    const rootKeyId = rootKey.value.key.id;
+    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
+    const rootKeyId = auth.key.id;
     await db.transaction(async (tx) => {
       await tx
         .update(schema.keys)
