+++ apps/api/src/routes/v1_apis_deleteApi.ts
@@ -1,7 +1,8 @@
-import { cache, db, keyService } from "@/pkg/global";
+import { cache, db } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 
+import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
 import { schema } from "@unkey/db";
 import { newId } from "@unkey/id";
@@ -50,20 +51,7 @@ export type V1ApisDeleteApiResponse = z.infer<
 
 export const registerV1ApisDeleteApi = (app: App) =>
   app.openapi(route, async (c) => {
-    const authorization = c.req.header("authorization")?.replace("Bearer ", "");
-    if (!authorization) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "key required" });
-    }
-    const rootKey = await keyService.verifyKey(c, { key: authorization });
-    if (rootKey.error) {
-      throw new UnkeyApiError({ code: "INTERNAL_SERVER_ERROR", message: rootKey.error.message });
-    }
-    if (!rootKey.value.valid) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "the root key is not valid" });
-    }
-    if (!rootKey.value.isRootKey) {
-      throw new UnkeyApiError({ code: "UNAUTHORIZED", message: "root key required" });
-    }
+    const auth = await rootKeyAuth(c);
 
     const { apiId } = c.req.valid("json");
 
@@ -75,11 +63,11 @@ export const registerV1ApisDeleteApi = (app: App) =>
       );
     });
 
-    if (!api || api.workspaceId !== rootKey.value.authorizedWorkspaceId) {
+    if (!api || api.workspaceId !== auth.authorizedWorkspaceId) {
       throw new UnkeyApiError({ code: "NOT_FOUND", message: `api ${apiId} not found` });
     }
-    const authorizedWorkspaceId = rootKey.value.authorizedWorkspaceId;
-    const rootKeyId = rootKey.value.key.id;
+    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
+    const rootKeyId = auth.key.id;
     await db.transaction(async (tx) => {
       await tx.update(schema.apis).set({ deletedAt: new Date() }).where(eq(schema.apis.id, apiId));
       await tx.insert(schema.auditLogs).values({
