+++ apps/api/src/routes/v1_keys_getKey.ts
@@ -1,7 +1,8 @@
-import { cache, db, keyService } from "@/pkg/global";
+import { cache, db } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 
+import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
 import { keySchema } from "./schema";
 
@@ -36,32 +37,7 @@ export type V1KeysGetKeyResponse = z.infer<
 >;
 export const registerV1KeysGetKey = (app: App) =>
   app.openapi(route, async (c) => {
-    const authorization = c.req.header("authorization")?.replace("Bearer ", "");
-    if (!authorization) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "key required",
-      });
-    }
-    const rootKey = await keyService.verifyKey(c, { key: authorization });
-    if (rootKey.error) {
-      throw new UnkeyApiError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: rootKey.error.message,
-      });
-    }
-    if (!rootKey.value.valid) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "the root key is not valid",
-      });
-    }
-    if (!rootKey.value.isRootKey) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "root key required",
-      });
-    }
+    const auth = await rootKeyAuth(c);
 
     const { keyId } = c.req.query();
 
@@ -88,7 +64,7 @@ export const registerV1KeysGetKey = (app: App) =>
       };
     });
 
-    if (!data || data.key.workspaceId !== rootKey.value.authorizedWorkspaceId) {
+    if (!data || data.key.workspaceId !== auth.authorizedWorkspaceId) {
       throw new UnkeyApiError({
         code: "NOT_FOUND",
         message: `key ${keyId} not found`,
