+++ apps/api/src/routes/v1_keys_getVerifications.ts
@@ -1,7 +1,8 @@
-import { analytics, cache, db, keyService } from "@/pkg/global";
+import { analytics, cache, db } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 
+import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
 
 const route = createRoute({
@@ -73,33 +74,9 @@ export type V1KeysGetVerificationsResponse = z.infer<
 >;
 export const registerV1KeysGetVerifications = (app: App) =>
   app.openapi(route, async (c) => {
-    const authorization = c.req.header("authorization")?.replace("Bearer ", "");
-    if (!authorization) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "key required",
-      });
-    }
-    const rootKey = await keyService.verifyKey(c, { key: authorization });
-    if (rootKey.error) {
-      throw new UnkeyApiError({
-        code: "INTERNAL_SERVER_ERROR",
-        message: rootKey.error.message,
-      });
-    }
-    if (!rootKey.value.valid) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "the root key is not valid",
-      });
-    }
-    if (!rootKey.value.isRootKey) {
-      throw new UnkeyApiError({
-        code: "UNAUTHORIZED",
-        message: "root key required",
-      });
-    }
-    const authorizedWorkspaceId = rootKey.value.authorizedWorkspaceId;
+    const auth = await rootKeyAuth(c);
+
+    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
     const { keyId, ownerId, start, end } = c.req.query();
 
     const ids: {
